Pushed more into AbstractView for reuse. Sped up scrolling significantly
