+++ enamlx/widgets/abstract_item_view.py
@@ -5,7 +5,7 @@ Created on Aug 23, 2015
 @author: jrm
 '''
 from atom.api import (
-    ContainerList, Enum, Bool, Property, ForwardInstance, observe
+    ContainerList, Int, Enum, Bool, Property, ForwardInstance, observe, set_default
 )
 from enaml.core.declarative import d_
 from enaml.widgets.control import Control, ProxyControl
@@ -29,8 +29,62 @@ class ProxyAbstractItemView(ProxyControl):
     
     def set_alternating_row_colors(self,alternate):
         pass
+    
+    def set_cell_padding(self,padding):
+        pass
+    
+    def set_auto_resize(self,enabled):
+        pass
+        
+    def set_resize_mode(self,mode):
+        pass
+    
+    def set_show_grid(self,show):
+        pass
+    
+    def set_word_wrap(self,enabled):
+        pass
+    
+    def set_show_vertical_header(self,visible):
+        pass
+    
+    def set_vertical_headers(self,headers):
+        pass
+    
+    def set_vertical_stretch(self,stretch):
+        pass
+    
+    def set_vertical_minimum_section_size(self,size):
+        pass
+    
+    def set_show_horizontal_header(self,visible):
+        pass
+    
+    def set_horizontal_headers(self,headers):
+        pass
+    
+    def set_horizontal_stretch(self,stretch):
+        pass
+    
+    def set_horizontal_minimum_section_size(self,size):
+        pass
+    
+    def set_sortable(self,sortable):
+        pass
+    
+    def set_visible_row(self,row):
+        pass
+    
+    def set_visible_column(self,column):
+        pass 
 
 class AbstractItemView(Control):
+    
+    #: Table should expand by default
+    hug_width = set_default('ignore')
+    
+    #: Table should expand by default
+    hug_height = set_default('ignore')
     
     #: The items to display in the view
     items = d_(ContainerList(default=[]))
@@ -47,6 +101,66 @@ class AbstractItemView(Control):
     #: Set alternating row colors
     alternating_row_colors = d_(Bool(False))
     
+    #: Cell padding
+    cell_padding = d_(Int(0))
+    
+    #: Automatically resize columns to fit contents
+    auto_resize = d_(Bool(True))
+    
+    #: Resize mode of columns and rows
+    resize_mode = d_(Enum('interactive','fixed','stretch','resize_to_contents','custom'))
+    
+    #: Show grid of cells
+    show_grid = d_(Bool(True))
+    
+    #: Word wrap
+    word_wrap = d_(Bool(False))
+    
+    #: Show vertical header bar
+    show_vertical_header = d_(Bool(True))
+    
+    #: Row headers 
+    vertical_headers = d_(ContainerList(basestring))
+    
+    #: Stretch last row
+    vertical_stretch = d_(Bool(False))
+    
+    #: Minimum row size
+    vertical_minimum_section_size = d_(Int(0))
+    
+    #: Show horizontal hearder bar
+    show_horizontal_header = d_(Bool(True))
+    
+    #: Column headers
+    horizontal_headers = d_(ContainerList(basestring))
+    
+    #: Stretch last column
+    horizontal_stretch = d_(Bool(False))
+    
+    #: Minimum column size
+    horizontal_minimum_section_size = d_(Int(0))
+    
+    #: Table is sortable
+    sortable = d_(Bool(True))
+    
+    #: Current row index
+    current_row = d_(Int(0))
+    
+    #: Current column index
+    current_column = d_(Int(0))
+    
+    #: First visible row
+    visible_row = d_(Int(0))
+    
+    #: Number of rows visible
+    visible_rows = d_(Int(100),writable=False)
+    
+    #: First visible column
+    visible_column = d_(Int(0))
+    
+    #: Number of columns visible
+    visible_columns = d_(Int(1),writable=False)
+    
     def _get_items(self):
         return [c for c in self.children if isinstance(c,AbstractWidgetItemGroup)]
     
@@ -54,7 +168,11 @@ class AbstractItemView(Control):
     _items = Property(_get_items,cached=True)
     
     @observe('items','scroll_to_bottom','alternating_row_colors',
-             'selection_mode','selection_behavior')
+             'selection_mode','selection_behavior','cell_padding',
+             'auto_resize','resize_mode','show_grid','word_wrap',
+             'show_horizontal_header','horizontal_headers','horizontal_stretch',
+             'show_vertical_header','vertical_header','vertical_stretch',
+             'visible_row','visible_column')
     def _update_proxy(self, change):
         """ An observer which sends state change to the proxy.
         """
